costMapCostToSBPLCost: never round down to 0
The previous logic didn't work: the division of two integers could
produce a 0, adding 0.5 and rounding down came back to 0. This could
lead to cells with a small cost value treated as free space, and
therefore the robot footprint would not get checked for collisions,
leading to invalid trajectories.

Thanks to Jorge Santos Simon (@corot)!